Use proper types when calling qBound.
qBound is a template function defined as:

 const T &qBound (const T &min, const T &value, const T &max )

On ARM, for some reason the type of min and max in this expression
resolves to double, and compiler fails to find a function matching
qBound(double, qreal, double). For that reason we typecast constants
to qreal explicitly.

+++ OpenRPT/renderer/satopaintengine.cpp
@@ -221,7 +221,7 @@ void SatoPaintEngine::drawBarcode ( const QPointF & p, const QString &format, in
   else if(format.contains("datamatrix"))
   {
     DmtxInfos dmtxInfos = extractInfosDtmx(format);
-    int eltSize = qRound (qBound(2.0, (qreal)height / (qreal)dmtxInfos.ySize, 20.0));
+    int eltSize = qRound (qBound((qreal)2.0, (qreal)height / (qreal)dmtxInfos.ySize, (qreal)20.0));
     barcodeFont = QString("BX0620%1%2%3%4")
         .arg(eltSize,2,10,QLatin1Char('0'))
         .arg(eltSize,2,10,QLatin1Char('0'))
